fix refactoring issues
