+++ src/databricks/labs/remorph/reconcile/query_builder/threshold_query.py
@@ -17,7 +17,7 @@ from databricks.labs.remorph.reconcile.query_builder.expression_generator import
     coalesce,
 )
 from databricks.labs.remorph.reconcile.recon_config import ColumnThresholds
-from databricks.labs.remorph.snow.databricks import Databricks
+from databricks.labs.remorph.transpiler.generator.databricks import Databricks
 
 logger = logging.getLogger(__name__)
