+++ src/databricks/labs/remorph/reconcile/schema_compare.py
@@ -12,7 +12,7 @@ from databricks.labs.remorph.reconcile.recon_config import (
     SchemaReconcileOutput,
     Table,
 )
-from databricks.labs.remorph.snow.databricks import Databricks
+from databricks.labs.remorph.transpiler.generator.databricks import Databricks
 
 logger = logging.getLogger(__name__)
