+++ tests/unit/transpiler/test_lca_utils.py
@@ -3,8 +3,8 @@ from unittest.mock import patch
 from sqlglot import parse_one
 
 from databricks.labs.remorph.config import get_dialect
-from databricks.labs.remorph.snow.databricks import Databricks
-from databricks.labs.remorph.snow.lca_utils import check_for_unsupported_lca
+from databricks.labs.remorph.transpiler.generator.databricks import Databricks
+from databricks.labs.remorph.transpiler.sqlglot.lca_utils import check_for_unsupported_lca
 
 
 def test_query_with_no_unsupported_lca_usage():
